refactor: organizando pastas
